08-nodered/node-red-voorbeeld-2/plc-sim/requirements.txt
asyncua
pymodbus
paho-mqtt